+++ qcio/constants.py
@@ -1,3 +1,17 @@
+"""Physical constants and periodic table from NIST CODATA 2022. All values should have a
+reference to their source.
+
+References:
+- https://physics.nist.gov/cuu/Constants/Table/allascii.txt
+- PubChem periodic table JSON: https://pubchem.ncbi.nlm.nih.gov/periodic-table/
+"""
+
+import csv
+from pathlib import Path
+from typing import Optional
+
+from pydantic import BaseModel
+
 # https://physics.nist.gov/cgi-bin/cuu/Value?bohrrada0
 BOHR_TO_ANGSTROM = 0.529177210544
 ANGSTROM_TO_BOHR = 1 / BOHR_TO_ANGSTROM
@@ -10,128 +24,84 @@ AVOGADRO_NUMBER = 6.02214076e23
 KCAL_TO_JOULE = 4.184e3
 HARTREE_TO_KCAL_PER_MOL = HARTREE_TO_JOULE / KCAL_TO_JOULE * AVOGADRO_NUMBER
 
-# Declaring Private so I can update implementations of this towards a periodic table.
-# Public features are exposed on core models. e.g., Structure.atomic_numbers
-_ELEMENTS = {
-    "H": 1,
-    "He": 2,
-    "Li": 3,
-    "Be": 4,
-    "B": 5,
-    "C": 6,
-    "N": 7,
-    "O": 8,
-    "F": 9,
-    "Ne": 10,
-    "Na": 11,
-    "Mg": 12,
-    "Al": 13,
-    "Si": 14,
-    "P": 15,
-    "S": 16,
-    "Cl": 17,
-    "Ar": 18,
-    "K": 19,
-    "Ca": 20,
-    "Sc": 21,
-    "Ti": 22,
-    "V": 23,
-    "Cr": 24,
-    "Mn": 25,
-    "Fe": 26,
-    "Co": 27,
-    "Ni": 28,
-    "Cu": 29,
-    "Zn": 30,
-    "Ga": 31,
-    "Ge": 32,
-    "As": 33,
-    "Se": 34,
-    "Br": 35,
-    "Kr": 36,
-    "Rb": 37,
-    "Sr": 38,
-    "Y": 39,
-    "Zr": 40,
-    "Nb": 41,
-    "Mo": 42,
-    "Tc": 43,
-    "Ru": 44,
-    "Rh": 45,
-    "Pd": 46,
-    "Ag": 47,
-    "Cd": 48,
-    "In": 49,
-    "Sn": 50,
-    "Sb": 51,
-    "Te": 52,
-    "I": 53,
-    "Xe": 54,
-    "Cs": 55,
-    "Ba": 56,
-    "La": 57,
-    "Ce": 58,
-    "Pr": 59,
-    "Nd": 60,
-    "Pm": 61,
-    "Sm": 62,
-    "Eu": 63,
-    "Gd": 64,
-    "Tb": 65,
-    "Dy": 66,
-    "Ho": 67,
-    "Er": 68,
-    "Tm": 69,
-    "Yb": 70,
-    "Lu": 71,
-    "Hf": 72,
-    "Ta": 73,
-    "W": 74,
-    "Re": 75,
-    "Os": 76,
-    "Ir": 77,
-    "Pt": 78,
-    "Au": 79,
-    "Hg": 80,
-    "Tl": 81,
-    "Pb": 82,
-    "Bi": 83,
-    "Po": 84,
-    "At": 85,
-    "Rn": 86,
-    "Fr": 87,
-    "Ra": 88,
-    "Ac": 89,
-    "Th": 90,
-    "Pa": 91,
-    "U": 92,
-    "Np": 93,
-    "Pu": 94,
-    "Am": 95,
-    "Cm": 96,
-    "Bk": 97,
-    "Cf": 98,
-    "Es": 99,
-    "Fm": 100,
-    "Md": 101,
-    "No": 102,
-    "Lr": 103,
-    "Rf": 104,
-    "Db": 105,
-    "Sg": 106,
-    "Bh": 107,
-    "Hs": 108,
-    "Mt": 109,
-    "Ds": 110,
-    "Rg": 111,
-    "Cn": 112,
-    "Nh": 113,
-    "Fl": 114,
-    "Mc": 115,
-    "Lv": 116,
-    "Ts": 117,
-    "Og": 118,
-}
+_DATA_DIR = Path(__file__).parent / "constants_data"
+
+
+class Atom(BaseModel):
+    """Atom data model."""
+
+    symbol: str
+    number: int
+    name: str
+    mass: float
+    group: Optional[int]
+    period: int
+    block: str
+    electron_config: str
+
+
+class PeriodicTable:
+    """Periodic table data model. Data source and URL should be set by class methods."""
+
+    # Should be set by @classmethod
+
+    data_source: str = ""
+    data_url: str = ""
+
+    @classmethod
+    def from_pubchem(cls):
+        """Create a periodic table from PubChem data."""
+        instance = cls()
+        instance.data_source = "PubChem"
+        instance.data_url = "https://pubchem.ncbi.nlm.nih.gov/periodic-table/"
+
+        group_and_period_data = {}
+
+        with open(_DATA_DIR / "group_period.csv", mode="r") as file:
+            reader = csv.DictReader(file)
+            for row in reader:
+                group_and_period_data[row["Symbol"]] = {
+                    "group": int(row["Group"]) if row["Group"] else None,
+                    "period": int(row["Period"]),
+                }
+
+        with open(_DATA_DIR / "pubchem.csv", mode="r") as file:
+            reader = csv.DictReader(file)
+            for row in reader:
+                symbol = row["Symbol"]
+                atom = Atom(
+                    symbol=symbol,
+                    number=int(row["AtomicNumber"]),
+                    name=row["Name"],
+                    mass=float(row["AtomicMass"]),
+                    group=group_and_period_data[symbol]["group"],
+                    period=group_and_period_data[symbol]["period"],
+                    block=row["GroupBlock"],
+                    electron_config=row["ElectronConfiguration"],
+                )
+                setattr(instance, symbol, atom)
+        return instance
+
+    def group(self, group_number: int):
+        """Return all atoms in a group."""
+        assert 1 <= group_number <= 18, "Group number must be between 1 and 18."
+        return [
+            atom
+            for atom in self.__dict__.values()
+            if isinstance(atom, Atom) and atom.group == group_number
+        ]
+
+    def period(self, period_number: int):
+        """Return all atoms in a period."""
+        assert 1 <= period_number <= 7, "Period number must be between 1 and 7."
+        return [
+            atom
+            for atom in self.__dict__.values()
+            if isinstance(atom, Atom) and atom.period == period_number
+        ]
+
+
+periodic_table = PeriodicTable.from_pubchem()
 
 # https://depts.washington.edu/eooptic/linkfiles/dielectric_chart%5B1%5D.pdf
 _SOLVENTS_DIELECTRIC = {